+++ types/object.d.ts
@@ -1,13 +1,16 @@
-export declare function add<T extends object>(
+export declare function addProperty<T extends object>(
 	object: T,
 	key: string,
 	value: any,
 ): T & Record<string, any>;
 export declare function deepCopy<T>(object: T): T;
 export declare function deepMerge<T extends object>(...objects: Partial<T>[]): T;
-export declare function forget<T extends object>(object: T, path: string): Partial<T>;
-export declare function get<T = null>(object: object, path: string, returnValue?: T): T;
-export declare function hasAny(object: object, paths: string[]): boolean;
+export declare function getPathValue<T = undefined>(
+	object: object,
+	path: string,
+	returnValue?: T,
+): T;
+export declare function hasAnyPath(object: object, paths: string[]): boolean;
 export declare function isNonEmptyObject(variable: any): variable is Record<string, any>;
 export declare function isObject(variable: any): variable is object;
 export declare function keyBy<T extends object>(array: T[], key: keyof T): Record<string, T>;
@@ -31,6 +34,7 @@ export declare function pick<T extends object, K extends keyof T>(
 	object: T,
 	properties: K[],
 ): Pick<T, K>;
-export declare function set<T extends object>(object: T, path: string, value: any): T;
+export declare function removePathValue<T extends object>(object: T, path: string): Partial<T>;
+export declare function setPathValue<T extends object>(object: T, path: string, value: any): T;
 export declare function unwrap(object: object): any | null;
 export declare function values(object: object): any[];